Put styling in JS file for optimization
To optimize nr of requests having to be made by the browser to use this
package, I put all styling into the javascript file. This way the
browser only needs to get one file instead of two (micro-optimize the
world!). This is a first step to a complete refactoring of this
directive, which will lead to a new major version.